Added another component for viewing forum posts
