+++ src/components/ForumPost.js
@@ -1,48 +1,13 @@
-import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Paper from '@material-ui/core/Paper';
-import Grid from '@material-ui/core/Grid';
-import Typography from '@material-ui/core/Typography';
+import React, { Component } from 'react';
 
-const useStyles = makeStyles(theme => ({
-  root: {
-    flexGrow: 1,
-  },
-  paper: {
-    padding: theme.spacing(2),
-    textAlign: 'center',
-    color: theme.palette.text.secondary,
-  },
-}));
+class ForumPost extends Component {
+    
+    render() {
+        return (
+            <h1>Here's where we'll view some forum posts!</h1>
+        )
+    }
 
-export default function ForumPost(props) {
-  const classes = useStyles();
-  let user = props.attributes.user
-  console.log(user)
-  return (
-    <div className={classes.root}>
-      <Grid container spacing={8}>
-        <Grid item xs={3}>
-          <Paper className={classes.paper}>
-            <Typography variant="h6" component="h6" gutterBottom>
-                {user.username}
-            </Typography>
-            {user.email}<br />
-            {user.first_name}
-            {user.last_name}
-          </Paper>
-        </Grid>
-        <Grid item xs={8}>
-          <Paper className={classes.paper} style={{textAlign: 'left'}}>
-            <Typography variant="h5" component="h5" gutterBottom>
-                {props.attributes.title}
-            </Typography>
-            <Typography variant="h6" component="h6" gutterBottom>
-                {props.attributes.description}
-            </Typography>
-          </Paper>
-        </Grid>
-      </Grid>
-    </div>
-  );
 }
+
+export default ForumPost
